students-assistant/pyproject.toml
[project]
name = "students-assitants"
version = "0.1.0"
description = "A Microsoft Teams chat bot based on Microsoft Teams echo bot"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "dotenv>=0.9.9",
  "discord.py>=2.4.0",
  "microsoft-teams-apps",
  "microsoft-teams-devtools"
]

[dependency-groups]
dev = [
    "pyright>=1.1.406",
]

[project.scripts]
start = "src.main:main"
typecheck = "pyright.cli:entrypoint"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
